Remove duplicate main code from HlmService.py, import from __main__.py
must: HLM_PV_Import/__main__.py This is exactly the same as the main function in HlmService.py why do we have it twice? Ideally there should only be one route into starting the program

Comments: Originally this was in order to have access to pv_import.stop() to be called on service stop. Now accessing pv_import explicitly on the __main__.py module so HlmService.py can use it too